Sycn with main

+++ coffee-kiosk-solution.Data/ViewModels/DiscountUpdateViewModel.cs
@@ -12,7 +12,7 @@ namespace coffee_kiosk_solution.Data.ViewModels
         [Required]
         public Guid Id { get; set; }
         [Required]
-        public float DiscountPercentage { get; set; }
+        public double DiscountPercentage { get; set; }
         [Required]
         public Guid ProductId { get; set; }
         [Required]
